Move softNavigate function into helper file
