+++ static/scripts/loggedin.js
@@ -1,50 +1,9 @@
 /* global kjua jQuery introJs*/
-import diffDom from 'diff-dom';
 import { setupFirebasePush } from './notificationService/indexFirebase';
 import { sendShownCallback, sendReadCallback} from './notificationService/callback';
 
 if (window.opener && window.opener !== window) {
     window.isInline = true;
-}
-
-const diffDOM = new diffDom();
-
-export function softNavigate(newurl, selector = 'html', listener, callback) {
-    $.ajax({
-        type: "GET",
-        url: newurl
-    }).done(function (r) {
-        // render new page
-        const parser = new DOMParser();
-        const newPage = parser.parseFromString(r, "text/html");
-        // apply new page
-        try {
-            const newPagePart = newPage.querySelector(selector);
-            const oldPagePart = document.querySelector(selector);
-            const diff = diffDOM.diff(oldPagePart, newPagePart);
-            const result = diffDOM.apply(oldPagePart, diff);
-            document.querySelectorAll((listener || selector) + " a").forEach(link => {
-                const linkClone = link.cloneNode(true);
-                linkClone.addEventListener("click", function (e) {
-                    softNavigate($(this).attr('href'), selector, listener);
-                    e.preventDefault();
-                    e.stopPropagation();
-                    return false;
-                });
-                link.parentNode.replaceChild(linkClone, link);
-            });
-            // scroll to top
-            document.body.scrollTop = 0; // For Safari
-            document.documentElement.scrollTop = 0; // For Chrome, Firefox, IE and Opera
-            jQuery(document).trigger('pageload');
-            if (callback) {
-                callback();
-            }
-        } catch (e) {
-            console.error(e);
-            $.showNotification("Fehler bei AJAX-Navigation", "danger", true);
-        }
-    });
 }
 
 function toggleMobileNav() {
@@ -66,7 +25,7 @@ function togglePresentationMode() {
     toggleButton.children('i').toggleClass('fa-expand');
 }
 
-var fullscreen = false;
+let fullscreen = false;
 
 function fullscreenBtnClicked() {
     togglePresentationMode();
